+++ packages/hmr-plugin/src/symbols.ts
@@ -23,9 +23,9 @@ export interface NgxsHmrLifeCycle<T = NgxsHmrSnapshot> {
 }
 
 export type HmrCallback<T> = (ctx: StateContext<T>, state: Partial<T>) => void;
-export type BootstrapModuleType<T> = () => Promise<NgModuleRef<T>>;
+export type BootstrapModuleFn<T = any> = () => Promise<NgModuleRef<T>>;
 
-export interface NgxsHmrOptions<T extends NgxsHmrLifeCycle<S>, S = NgxsHmrSnapshot> {
+export interface NgxsHmrOptions {
   /**
    * @description
    * clear log after each hmr update
@@ -53,4 +53,4 @@ interface HotModule {
  * any - because need setup
  * npm i @types/webpack-env
  */
-export type WebpackModule = (NodeModule & HotModule) | any;
+export type WebpackModule = HotModule | any;