+++ packages/logger-plugin/tests/helpers/index.ts
@@ -1,0 +1,2 @@
+export * from './logger-spy';
+export * from './utils';