+++ packages/router-plugin/src/router.module.ts
@@ -3,13 +3,13 @@ import { NgxsModule } from '@ngxs/store';
 import { DefaultRouterStateSerializer, RouterStateSerializer } from './serializer';
 import { RouterState } from './router.state';
 
-// NOTE: Must mark as `dynamic` due to
-// https://github.com/dherges/ng-packagr/issues/767
-export const NgxsModuleRouterState = NgxsModule.forFeature([RouterState]);
-
 // @dynamic
 @NgModule({
-  imports: [NgxsModuleRouterState],
+  imports: [
+    // NOTE: Must mark as `dynamic` due to
+    // https://github.com/dherges/ng-packagr/issues/767
+    NgxsModule.forFeature([RouterState])
+  ],
   providers: [{ provide: RouterStateSerializer, useClass: DefaultRouterStateSerializer }]
 })
 export class NgxsRouterPluginModule {