+++ packages/store/operators/src/index.ts
@@ -1,3 +1,8 @@
+/**
+ * @module
+ * @description
+ * Entry point for all public APIs of this package.
+ */
 export { append } from './append';
 export { compose } from './compose';
 export { iif } from './iif';
@@ -5,3 +10,4 @@ export { insertItem } from './insert-item';
 export { patch } from './patch';
 export { isStateOperator } from './utils';
 export { updateItem } from './update-item';
+export { removeItem } from './remove-item';