+++ packages/store/tests/zone.spec.ts
@@ -19,7 +19,7 @@ import { Observable, Subscription } from 'rxjs';
 import { take } from 'rxjs/operators';
 
 import { State, Action, StateContext, NgxsModule, Store, Select } from '../src/public_api';
-import { NoopNgxsExecutionStrategy } from '../src/execution/noopNgxsExecutionStrategy';
+import { NoopNgxsExecutionStrategy } from '../src/execution/noop-ngxs-execution-strategy';
 
 describe('zone', () => {
   class Increment {
@@ -189,7 +189,6 @@ describe('zone', () => {
     const store: Store = TestBed.get(Store);
     const ngZone: NgZone = TestBed.get(NgZone);
     ngZone.run(() => {
-      console.log({ isInAngularZone: NgZone.isInAngularZone() });
       store.dispatch(new FooAction());
     });
   });